feat(06-07): fix CurrentGauges decode for exponent-form Java doubles
- heap/max gauges decode as f64: 8.3.3 (b2026012009) serializes Java
  doubles in scientific notation (heapMemory: 2.85746728E8) which
  serde_json refuses for i64
- serialize_bytes_f64 keeps whole byte counts as JSON integers so
  agent-visible --json output keeps the pre-f64 shape
- exponent/integer/decimal-form fixture pinned via from_str (json!
  pre-normalizes numbers and cannot prove exponent TEXT decodes)
- sibling audit documented: the two memory gauges are the only
  double-typed gauge fields on the wire (threads are longs, chart
  values already f64)
- display consumers: render.rs casts at call site, fmt_mib takes f64 —
  rendered output byte-identical for whole bytes, goldens stay green

Closes 06-UAT test 4 [major]

+++ crates/ignition-cli/src/render.rs
@@ -360,8 +360,10 @@ fn render_metrics_human(result: &MetricsResult) {
     println!(
         "cpu {:.1}%  heap {}/{}",
         gauges.cpu,
-        human_bytes(gauges.heap_memory),
-        human_bytes(gauges.max_memory)
+        // f64 gauges (8.3.3 exponent form); whole byte counts cast
+        // exact (≤2^53) — the datapoint.value precedent below.
+        human_bytes(gauges.heap_memory as i64),
+        human_bytes(gauges.max_memory as i64)
     );
     let threads = &result.threads;
     println!(
